refactor: View 파일 분리, UITextView Extension추가
